+++ src/main/java/net/sf/mzmine/modules/peaklistmethods/identification/spectraldbsearch/LocalSpectralDBSearchModule.java
@@ -26,8 +26,8 @@ import net.sf.mzmine.datamodel.PeakListRow;
 import net.sf.mzmine.main.MZmineCore;
 import net.sf.mzmine.modules.MZmineModuleCategory;
 import net.sf.mzmine.modules.MZmineProcessingModule;
-import net.sf.mzmine.modules.peaklistmethods.identification.spectraldbsearch.selectedrows.singlerows.SingleRowLocalSpectralDBSearchParameters;
-import net.sf.mzmine.modules.peaklistmethods.identification.spectraldbsearch.selectedrows.singlerows.SingleRowLocalSpectralDBSearchTask;
+import net.sf.mzmine.modules.peaklistmethods.identification.spectraldbsearch.selectedrows.singlerow.SingleRowLocalSpectralDBSearchParameters;
+import net.sf.mzmine.modules.peaklistmethods.identification.spectraldbsearch.selectedrows.singlerow.SingleRowLocalSpectralDBSearchTask;
 import net.sf.mzmine.modules.visualization.peaklisttable.table.PeakListTable;
 import net.sf.mzmine.parameters.ParameterSet;
 import net.sf.mzmine.taskcontrol.Task;
