+++ src/main/java/net/sf/mzmine/modules/peaklistmethods/identification/spectraldbsearch/selectedrows/singlerow/SingleRowLocalSpectralDBSearchParameters.java
@@ -1,4 +1,4 @@
-package net.sf.mzmine.modules.peaklistmethods.identification.spectraldbsearch.selectedrows.singlerows;
+package net.sf.mzmine.modules.peaklistmethods.identification.spectraldbsearch.selectedrows.singlerow;
 
 import java.awt.Window;
 import javax.swing.JComponent;
